Temporarily point federation discovery at dev instances
There currently isn't a production instance of an osdf registry or director set up. To keep things
working temporarily, this points at the dev instances (which should be working). Anyone who registers
an origin using the new CLI might have to re-register when we point this someplace new, but that's a
pretty easy thing to do at this point.